Fix: 전역 가드 오류 수정

+++ src/users/users.service.ts
@@ -3,6 +3,7 @@ import { UpdateUserDto } from "./dto/update-user.dto";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Users } from "./entities/user.entity";
 import { Repository } from "typeorm";
+import { User } from "src/common/decorator/user.decorator";
 
 @Injectable()
 export class UsersService {
@@ -20,15 +21,25 @@ export class UsersService {
     if(!user){
       throw new Error("유저가 존재하지 않습니다.");
     }
+
+    if(user.IsVaildated === false){
+      throw new Error("이메일 인증을 거쳐야 하는 회원입니다.");
+    }
+
     return user;
   }
 
   async update(userId : number, updateUserDto: UpdateUserDto) {
+    const user = await this.findid(userId);
 
-    if(!userId){
+    if(!user){
       throw new Error("유저가 존재하지 않습니다.");
     }
-
+    
+    if(user.IsVaildated === false){
+      throw new Error("이메일 인증을 거쳐야 하는 회원입니다.");
+    }
+  
     return await this.userRepository.update(userId, updateUserDto);
   }
 
@@ -48,6 +59,16 @@ export class UsersService {
   }
 
   async remove(userId : number) {
+    const user = await this.findid(userId);
+
+    if(!userId){
+      throw new Error("유저가 존재하지 않습니다.");
+    }
+
+    if(user.IsVaildated === false){
+      throw new Error("이메일 인증을 거쳐야 하는 회원입니다.");
+    }
+
     return await this.userRepository.delete(userId);
   }
